- Changed from entity / component model to a node based one nodes have children of type node. This seems to improve performance significantly and removes bloat that isn't needed for our purposes

+++ ChaosEngine/src/Chaos/Core/Time.cpp
@@ -5,7 +5,7 @@ namespace Chaos
 {
 	float Time::m_deltaTime = 0.0f;
 	float Time::m_fixedDeltaTime = 0.22f;
-	float Time::m_time = 0.0f;
-	float Time::m_timeLastFrame = 0.0f;
+	double Time::m_time = 0.0f;
+	double Time::m_timeLastFrame = 0.0f;
 	float Time::m_timeScale = 1.0f;
 }
